% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCoords.R
\name{generateCoords}
\alias{generateCoords}
\title{Generates the coordinates for the plot of a sequence object based on a precomputed layout}
\usage{
generateCoords(seqObj, layout)
}
\arguments{
\item{seqObj}{the sequence object}

\item{layout}{layout computed with generateLayout()}
}
\value{
data.table object
}
\description{
Generates the coordinates for the plot of a sequence object based on a precomputed layout
}
