% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateLayout.R
\name{generateLayout}
\alias{generateLayout}
\title{Generates the layout for the plot of a sequence object}
\usage{
generateLayout(seqObj)
}
\arguments{
\item{seqObj}{the sequence object}
}
\value{
data.frame object
}
\description{
Generates the layout for the plot of a sequence object
}
