% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectoryplot.R
\name{trajectoryplot}
\alias{trajectoryplot}
\title{Create a new sequence active plot}
\usage{
trajectoryplot(
  seqdata,
  alphabet = NULL,
  names = NULL,
  layout = NULL,
  ylab = NULL,
  yaxis = TRUE,
  strokeStyle = "rgb(255,0,0,0.1)",
  groupColor = NULL,
  group = NULL,
  area = NULL,
  width = "100\%",
  height = 1000,
  margins = c(20, 20, 20, 20),
  padding = list(top = 10, bottom = 10, left = 50, right = 10),
  innerPadding = c(0, 0, 0, 0),
  xlabel = NULL,
  position = "center",
  alpha = NULL,
  ...
)
}
\arguments{
\item{seqdata}{a state sequence object or a \code{generateCoords(seqobj)} generated coordinates with a passed layout (see layout).}

\item{names}{column names of the sequence object}

\item{layout}{A layout that provides the plot layout. Can be used to 1) draw a global layout where 2) local sequences (e.g. by group) are drawn.}

\item{ylab}{Labels for states}

\item{strokeStyle}{global stroke style. Any valid html color specification is valid. user rgba() for transparency e.g. \code{color="rgba(255,0,0,0.5)"}}

\item{groupColor}{List of html color specification to color factor levels of group.}

\item{group}{Varialbe whos factor levels are used for coloring the plot. Requieres a list of groupColors.}

\item{margins}{determines the plot's margin, starting with "top" and going clockwise: \code{c(top, right, bottom, left)}. Defaults to \code{c(20,20,20,20)}.}
}
\value{
trajectoryplot object
}
\description{
\code{seqaplot()} initializes the sequence active plot html widget.
By default, it returns a static plot that can be enhanced with adding interactive functions.
}
